feature(web): Context menu -> edit priority, refactor code
- Implement context menu edit priority
- Decouple `ContextMenu` and create `ContextMenuItems` to separate floating ui menu logic from menu items business logic
